Move problem report handling from supportPage
To be able to handle problem reports on mobile. Also moves ipc for collect-logs from tray to onReady.
